Updated the branch with the latest changes from the main trunk (revision 1699-1712): - Fixed: When using space as jump key, using the in game join menu made the player jump continuously after spawning. This bug was caused by the changes from revision 1444. - Fixed: The scoreboard of the win32 server console is not updated when changing to intermission. - Fixed: The game crashed while spawning a player in certain game modes (e.g. lms), when certain inventory items (e.g. weapons) were replaced by actors that are not inventory items. - Removed the "jump key" history entry since that bug was not present in 97d2. - Improved QueryIPQueue. [Rivecoder]     - Reverted to the fixed-sized array to regain O(1) efficiency in addAddress.     - Allowed each queue to have a specific entry length, set via its constructor. Replaced all queues that used mixed entry lengths with multiple queues.     - Moved to networkshared.cpp so it can be edited without triggering a mass recompile.     - Fixed a bug where if the queue was filled, it wouldn't remove elements correctly (or even get stuck).     - Added isFull().     - Simplified some of the looping code. - If the flood queue is full, the server will reject new connections. This allows players to keep playing even if the server is being attacked. [Rivecoder] - Cleaned up networkshared.h; added better dividers, tabs, and proper spacing. [Rivecoder] - Fixed a rare assertion bug that triggered in 640x480 debug mode. It was actually introduced in revision 153. [Rivecoder] - Fixed: When entering the name of the save game in the save menu, the underscore denoting the cursor position is possibly displayed too far to the right. - Removed member variable _iNumberOfEntries from QueryIPQueue. Keeping track of the number of entries is not necessary, because it can be directly calculated from _iQueueHead and _iQueueTail. Furthermore, iNumberOfEntries was wrong as soon as addAddress was called when the queue is full. - The addition of the "in console" icon in revision 1442 messed up the sprite numbers used in medal_SelectIcon for ulActualSprite and ulDesiredSprite. The ally icon had ulActualSprite == 1 (already used by the "in console" icon) and ulDesiredSprite == 2 (already used by the flag / skull icons). To fix this kind of problems once and for all, I added a new enumerator. The whole approach of hard coding the numbers at different places was designated to break someday.
SVN r1742 (latestzdoom)

+++ src/m_menu.cpp
@@ -1360,7 +1360,8 @@ static void M_DrawSaveLoadCommon ()
 						listboxTop+rowHeight*i+CleanYfac, savegamestring,
 						DTA_CleanNoMove, true, TAG_DONE);
 					screen->DrawText (CR_WHITE,
-						listboxLeft+1+SmallFont->StringWidth (savegamestring)*CleanXfac,
+						// [BB] Using Skulltag's increased precision of CleanXfac is wrong here.
+						listboxLeft+1+SmallFont->StringWidth (savegamestring)*static_cast<int>(CleanXfac),
 						listboxTop+rowHeight*i+CleanYfac, underscore,
 						DTA_CleanNoMove, true, TAG_DONE);
 				}
